Refactor: refactor map page 맵 페이지 리팩토링
