+++ alcoholmap/src/frontend/src/components/units/map/maphbmenu/MapHBMenu.tsx
@@ -5,14 +5,6 @@ import { useRouter } from "next/router";
 export default function MapHBMenu (props: IHBMenuProps): JSX.Element {
     const router = useRouter();
 
-    const onClickMoveToKorea = () => {
-        router.push('/map/korea')
-    };
-
-    const onClickMoveToSeoul = () => {
-        router.push('/map/seoul')
-    };
-
     return(
         <>
             {props.isHBMenuOpen && (
@@ -32,8 +24,8 @@ export default function MapHBMenu (props: IHBMenuProps): JSX.Element {
                         </S.MyInfoWrapper>
                         <S.Line/>
                         <S.HBMenuList>
-                            <S.HBMenuListEl onClick={onClickMoveToKorea}>지역별 술집 추천 (팔도)</S.HBMenuListEl>
-                            <S.HBMenuListEl onClick={onClickMoveToSeoul}>지역별 술집 추천 (서울)</S.HBMenuListEl>
+                            <S.HBMenuListEl onClick={props.onClickMoveToKorea}>지역별 술집 추천 (팔도)</S.HBMenuListEl>
+                            <S.HBMenuListEl onClick={props.onClickMoveToSeoul}>지역별 술집 추천 (서울)</S.HBMenuListEl>
                         </S.HBMenuList>
                     </S.HBMenuWrapper>
             )}
